Cleanup, restructuring files & bottom sheet bugfix
