+++ architecture-compose/src/androidLibMain/kotlin/com/splendo/kaluga/architecture/compose/navigation/BottomSheetRoute.kt
@@ -1,0 +1,47 @@
+/*
+ Copyright 2022 Splendo Consulting B.V. The Netherlands
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
+package com.splendo.kaluga.architecture.compose.navigation
+
+import androidx.compose.material.ModalBottomSheetLayout
+
+/**
+ * Route for navigating a [ModalBottomSheetLayout] using a [BottomSheetSheetContentRouteController]
+ */
+sealed class BottomSheetRoute {
+    /**
+     * Navigates within the sheet content of the [ModalBottomSheetLayout]
+     * @param route The [Route] to navigate
+     */
+    data class SheetContent(val route: Route) : BottomSheetRoute()
+
+    /**
+     * Navigates within the content of the [ModalBottomSheetLayout]
+     * @param route The [Route] to navigate
+     */
+    data class Content(val route: Route) : BottomSheetRoute()
+}
+
+/**
+ * Converts a [Route] to a [BottomSheetRoute.SheetContent] route
+ */
+val Route.bottomSheetSheetContent get() = BottomSheetRoute.SheetContent(this)
+
+/**
+ * Converts a [Route] to a [BottomSheetRoute.Content] route
+ */
+val Route.bottomSheetContent get() = BottomSheetRoute.Content(this)
