+++ architecture-compose/src/androidLibMain/kotlin/com/splendo/kaluga/architecture/compose/navigation/BottomSheetRouteController.kt
@@ -1,0 +1,33 @@
+/*
+ Copyright 2022 Splendo Consulting B.V. The Netherlands
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
+package com.splendo.kaluga.architecture.compose.navigation
+
+import androidx.compose.material.ModalBottomSheetLayout
+
+/**
+ * A controller that handles the [Route] for a [ModalBottomSheetLayout]
+ */
+class BottomSheetRouteController(
+    internal val contentRouteController: RouteController,
+    internal val sheetContentRouteController: BottomSheetSheetContentRouteController
+) {
+    fun navigate(route: BottomSheetRoute) = when (route) {
+        is BottomSheetRoute.SheetContent -> sheetContentRouteController.navigate(route.route)
+        is BottomSheetRoute.Content -> contentRouteController.navigate(route.route)
+    }
+}
